feat(useSearchWithRouter.tsx): improve chip styling and add capitalization
This commit improves the styling of the chips in the `useSearchWithRouter`
hook. The size of the chips is changed to "medium", the margin is increased
to "5px", and the `capitalizeFirstCharacter` function is used to capitalize
the first character of the chip label. This makes the chips more visually
appealing and easier to read.