feat(genome-page): display NoMatchDisplay component when no publications are found
This change introduces a `NoMatchDisplay` component that is rendered when the `filteredAndSorted` array is empty. This provides a better user experience by informing the user that no publications match the current filter criteria. The `fp-ts` library is used to handle the conditional rendering of either the `NoMatchDisplay` component or the list of `PublicationRow` components.