feat(genome-page): improve formatting of references in table
This commit improves the formatting of references in the
ReferencesPanel component. It adds a period after the authors'
last names, removes unnecessary spaces, and adds a comma after
the journal name. It also introduces a new style for the
reference cell to add padding. These changes improve the
readability and visual appeal of the references.