feat(genome-page): create RelatedGenesList component
This component displays a list of related genes for a given
publication. It takes a publicationId, a limit for the number
of genes to display, and an array of gene objects as props.
It uses Material-UI components for styling and layout, and
fp-ts for functional programming. Each gene is displayed as
a clickable chip that navigates to the gene's page when clicked.
If the number of genes exceeds the limit, a "See All Genes"
chip is displayed, which navigates to a page displaying all
related genes for the publication.