test(GoaPanel.test.tsx): update test to check for no GO sections when no GOAs are present
The test case was updated to check that no GO sections are rendered when the goas prop is an empty array. The previous test case was filtering the goas prop to remove molecular_function, but this was not a good test case because it was still rendering the other two GO sections. The updated test case is more accurate because it checks that no GO sections are rendered when there are no GOAs present.