feat(generate-dependency-graph): add --local-only flag to filter out non-local dependencies
This change introduces a new `--local-only` flag to the
`generate-dependency-graph` package. When this flag is used, the
dependency tree will only include local dependencies (i.e., those
found within the `localPackages` record). This is achieved by
filtering the dependencies object before generating the dependency
tree. A global variable `localOnly` is used to pass the flag value
to the `getDependencyTree2` function.