feat: add oxlint script to all packages
This change adds a new script called "oxlint" to the package.json
file of every package. The script runs the oxlint linter on the src/
directory of the package. Additionally, an "oxlint:out" script is
added to output the lint results in GitHub format. This change
enables the use of oxlint for linting the code in each package,
providing an alternative to eslint.