refactor(listStrainData.ts): use Array.from to generate array
The change replaces the `new Array(length).fill(0)` method of
generating an array with `Array.from({ length })`. This is done
because `new Array(length).fill(0)` can have unexpected behavior
when used with objects, as it creates an array where each element
references the same object. `Array.from({ length })` avoids this
issue by creating a new object for each element in the array.