feat(.oxlintrc.json): configure oxlint to remove jest and add react version detect
The jest plugin is removed from the oxlint configuration as it is not used in the project.
The react version is set to "detect" in the settings to ensure that oxlint uses the correct version of React when linting React components.
The no-standalone-expect rule is disabled to allow expect statements to be used outside of test blocks.