refactor: migrate to ui-common-mui5 package
This commit migrates the application to use the `@dictybase/ui-common-mui5` package instead of `@dictybase/ui-common`. This change ensures that the application uses the latest version of the dictyBase UI components, which are now based on Material UI v5.
The ui-user package was also updated to the latest version.

The ui-common package is deprecated and will no longer be maintained.