refactor(dicty-frontpage-mui5): replace div elements with Box component from Material-UI
This commit refactors the code to replace `div` elements with the `Box` component from Material-UI. This change improves the semantics and consistency of the code, as `Box` is a more appropriate component for layout and styling purposes within a Material-UI application. It also allows for easier customization and theming of the components.