feat(genome-page-mui5): rename app and update dependencies
This commit renames the app from `genomepage` to `genomepage-mui5`
to reflect the migration to Material UI v5. It also updates the
dependencies to use the MUI5 compatible `@dictybase/auth-mui5`
package and adds `@dictybase/ui-common`. The lint output path is
updated to reflect the new app name. Finally, the commit removes
the deprecated `@material-ui/*` packages.

The renaming and dependency updates ensure that the app is
correctly configured for MUI5 and uses the appropriate packages.