refactor(auth-mui5): migrate from `@mui/styles` to `tss-react/mui` for styling
Migrates the styling solution from the deprecated `@mui/styles` to
`tss-react/mui` to ensure compatibility with newer versions of Material
UI and improve maintainability. This involves replacing `makeStyles`
from `@mui/styles` with `makeStyles` from `tss-react/mui` and
adjusting the usage of the resulting `classes` object in the components.

The `@mui/styles` package is deprecated and no longer recommended for
use in new projects. `tss-react/mui` is a modern styling solution that
is compatible with newer versions of Material UI and offers improved
performance and maintainability.